feat: update delete logic

+++ image_modules/main.py
@@ -7,6 +7,7 @@ def get_argument():
     parser.add_argument("--url", required=False, help="")
     parser.add_argument("--src_dir", required=False, help="")
     parser.add_argument("--compressed_dir", required=False, help="")
+    parser.add_argument("--delete_dir", required=False, help="")
     
     return parser.parse_args()
 
@@ -17,9 +18,12 @@ if __name__ == "__main__":
     src_dir = args.src_dir
 
     url = args.url
-    downloader.delete_all_files_in_directory(src_dir)
-    downloader.download_images(url, src_dir)
-    
     compressed_dir = args.compressed_dir
-    downloader.delete_all_files_in_directory(compressed_dir)    
+    if args.delete_dir:
+        downloader.delete_all_files_in_directory(src_dir)
+        downloader.delete_all_files_in_directory(compressed_dir)    
+
+
+    downloader.download_images(url, src_dir)
+
     compress_image.resize_images_in_directory(src_dir,compressed_dir)
